+++ sdk/search/search-documents/src/searchClient.ts
@@ -8,10 +8,12 @@ import {
   InternalPipelineOptions,
   createPipelineFromOptions,
   OperationOptions,
-  operationOptionsToRequestOptionsBase
+  operationOptionsToRequestOptionsBase,
+  RequestPolicyFactory,
+  bearerTokenAuthenticationPolicy,
 } from "@azure/core-http";
 import { SearchClient as GeneratedClient } from "./generated/data/searchClient";
-import { KeyCredential } from "@azure/core-auth";
+import { KeyCredential, TokenCredential, isTokenCredential } from "@azure/core-auth";
 import { createSearchApiKeyCredentialPolicy } from "./searchApiKeyCredentialPolicy";
 import { SDK_VERSION } from "./constants";
 import { logger } from "./logger";
@@ -19,7 +21,7 @@ import {
   AutocompleteResult,
   AutocompleteRequest,
   SuggestRequest,
-  IndexDocumentsResult
+  IndexDocumentsResult,
 } from "./generated/data/models";
 import { createSpan } from "./tracing";
 import { SpanStatusCode } from "@azure/core-tracing";
@@ -41,13 +43,15 @@ import {
   DeleteDocumentsOptions,
   SearchDocumentsPageResult,
   MergeOrUploadDocumentsOptions,
-  SearchRequest
+  SearchRequest,
 } from "./indexModels";
 import { odataMetadataPolicy } from "./odataMetadataPolicy";
 import { IndexDocumentsBatch } from "./indexDocumentsBatch";
 import { encode, decode } from "./base64";
 import * as utils from "./serviceUtils";
 import { IndexDocumentsClient } from "./searchIndexingBufferedSender";
+import { KnownSearchAudience } from "./searchAudience";
+
 /**
  * Client options used to configure Cognitive Search API requests.
  */
@@ -56,6 +60,13 @@ export interface SearchClientOptions extends PipelineOptions {
    * The API version to use when communicating with the service.
    */
   apiVersion?: string;
+
+  /**
+   * The Audience to use for authentication with Azure Active Directory (AAD). The
+   * audience is not considered when using a shared key.
+   * {@link KnownSearchAudience} can be used interchangeably with audience
+   */
+  audience?: string;
 }
 
 /**
@@ -110,7 +121,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
   constructor(
     endpoint: string,
     indexName: string,
-    credential: KeyCredential,
+    credential: KeyCredential | TokenCredential,
     options: SearchClientOptions = {}
   ) {
     this.endpoint = endpoint;
@@ -137,16 +148,22 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
             "OData-MaxVersion",
             "OData-Version",
             "Prefer",
-            "throttle-reason"
-          ]
-        }
-      }
+            "throttle-reason",
+          ],
+        },
+      },
     };
 
-    const pipeline = createPipelineFromOptions(
-      internalPipelineOptions,
-      createSearchApiKeyCredentialPolicy(credential)
-    );
+    const scope: string = options.audience
+      ? `${options.audience}/.default`
+      : `${KnownSearchAudience.AzurePublicCloud}/.default`;
+
+    const requestPolicyFactory: RequestPolicyFactory = isTokenCredential(credential)
+      ? bearerTokenAuthenticationPolicy(credential, scope)
+      : createSearchApiKeyCredentialPolicy(credential);
+
+    const pipeline = createPipelineFromOptions(internalPipelineOptions, requestPolicyFactory);
+
     if (Array.isArray(pipeline.requestPolicyFactories)) {
       pipeline.requestPolicyFactories.unshift(odataMetadataPolicy("none"));
     }
@@ -154,7 +171,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     let apiVersion = this.apiVersion;
 
     if (options.apiVersion) {
-      if (!["2020-06-30-Preview", "2020-06-30"].includes(options.apiVersion)) {
+      if (!["2020-06-30", "2021-04-30-Preview"].includes(options.apiVersion)) {
         throw new Error(`Invalid Api Version: ${options.apiVersion}`);
       }
       apiVersion = options.apiVersion;
@@ -177,7 +194,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -203,7 +220,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
       searchText: searchText,
       suggesterName: suggesterName,
       searchFields: this.convertSearchFields<Fields>(searchFields),
-      ...nonFieldOptions
+      ...nonFieldOptions,
     };
 
     if (!fullOptions.searchText) {
@@ -225,7 +242,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -245,7 +262,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
       select: this.convertSelect<Fields>(select),
       orderBy: this.convertOrderBy(orderBy),
       ...nonFieldOptions,
-      ...nextPageParameters
+      ...nextPageParameters,
     };
 
     const { span, updatedOptions } = createSpan("SearchClient-searchDocuments", operationOptions);
@@ -255,7 +272,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
         {
           ...fullOptions,
           includeTotalResultCount: fullOptions.includeTotalCount,
-          searchText: searchText
+          searchText: searchText,
         },
         operationOptionsToRequestOptionsBase(updatedOptions)
       );
@@ -270,14 +287,14 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
         coverage,
         facets,
         answers,
-        continuationToken: this.encodeContinuationToken(nextLink, result.nextPageParameters)
+        continuationToken: this.encodeContinuationToken(nextLink, result.nextPageParameters),
       };
 
       return deserialize<SearchDocumentsPageResult<Pick<T, Fields>>>(converted);
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -320,7 +337,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     yield* firstPage.results;
     if (firstPage.continuationToken) {
       for await (const page of this.listSearchResultsPage(searchText, options, {
-        continuationToken: firstPage.continuationToken
+        continuationToken: firstPage.continuationToken,
       })) {
         yield* page.results;
       }
@@ -343,7 +360,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
       },
       byPage: (settings: ListSearchResultsPageSettings = {}) => {
         return this.listSearchResultsPage(searchText, options, settings);
-      }
+      },
     };
   }
 
@@ -369,12 +386,12 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
         coverage,
         facets,
         answers,
-        results: this.listSearchResults(pageResult, searchText, updatedOptions)
+        results: this.listSearchResults(pageResult, searchText, updatedOptions),
       };
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -402,7 +419,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
       searchFields: this.convertSearchFields<Fields>(searchFields),
       select: this.convertSelect<Fields>(select),
       orderBy: this.convertOrderBy(orderBy),
-      ...nonFieldOptions
+      ...nonFieldOptions,
     };
 
     if (!fullOptions.searchText) {
@@ -421,15 +438,14 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
         operationOptionsToRequestOptionsBase(updatedOptions)
       );
 
-      const modifiedResult = utils.generatedSuggestDocumentsResultToPublicSuggestDocumentsResult<T>(
-        result
-      );
+      const modifiedResult =
+        utils.generatedSuggestDocumentsResultToPublicSuggestDocumentsResult<T>(result);
 
       return deserialize<SuggestDocumentsResult<Pick<T, Fields>>>(modifiedResult);
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -456,7 +472,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -492,7 +508,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -519,7 +535,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -547,7 +563,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -575,7 +591,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -624,7 +640,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -642,7 +658,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
     const payload = JSON.stringify({
       apiVersion: this.apiVersion,
       nextLink,
-      nextPageParameters
+      nextPageParameters,
     });
     return encode(payload);
   }
@@ -669,7 +685,7 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
 
       return {
         nextLink: result.nextLink,
-        nextPageParameters: result.nextPageParameters
+        nextPageParameters: result.nextPageParameters,
       };
     } catch (e) {
       throw new Error(`Corrupted or invalid continuation token: ${decodedToken}`);
@@ -689,9 +705,9 @@ export class SearchClient<T> implements IndexDocumentsClient<T> {
       operationOptions: {
         abortSignal,
         requestOptions,
-        tracingOptions
+        tracingOptions,
       },
-      restOptions
+      restOptions,
     };
   }