+++ sdk/search/search-documents/src/searchIndexClient.ts
@@ -3,12 +3,14 @@
 
 /// <reference lib="esnext.asynciterable" />
 
-import { KeyCredential } from "@azure/core-auth";
+import { KeyCredential, TokenCredential, isTokenCredential } from "@azure/core-auth";
 import {
   createPipelineFromOptions,
   InternalPipelineOptions,
   operationOptionsToRequestOptionsBase,
-  PipelineOptions
+  PipelineOptions,
+  RequestPolicyFactory,
+  bearerTokenAuthenticationPolicy,
 } from "@azure/core-http";
 import { SpanStatusCode } from "@azure/core-tracing";
 import { SDK_VERSION } from "./constants";
@@ -35,12 +37,13 @@ import {
   IndexIterator,
   IndexNameIterator,
   SearchIndexStatistics,
-  SearchServiceStatistics
+  SearchServiceStatistics,
 } from "./serviceModels";
 import * as utils from "./serviceUtils";
 import { createSpan } from "./tracing";
 import { odataMetadataPolicy } from "./odataMetadataPolicy";
 import { SearchClient, SearchClientOptions as GetSearchClientOptions } from "./searchClient";
+import { KnownSearchAudience } from "./searchAudience";
 
 /**
  * Client options used to configure Cognitive Search API requests.
@@ -50,6 +53,13 @@ export interface SearchIndexClientOptions extends PipelineOptions {
    * The API version to use when communicating with the service.
    */
   apiVersion?: string;
+
+  /**
+   * The Audience to use for authentication with Azure Active Directory (AAD). The
+   * audience is not considered when using a shared key.
+   * {@link KnownSearchAudience} can be used interchangeably with audience
+   */
+  audience?: string;
 }
 
 /**
@@ -78,7 +88,7 @@ export class SearchIndexClient {
   /**
    * Used to authenticate requests to the service.
    */
-  private readonly credential: KeyCredential;
+  private readonly credential: KeyCredential | TokenCredential;
 
   /**
    * Used to configure the Search Index client.
@@ -101,7 +111,11 @@ export class SearchIndexClient {
    * @param credential - Used to authenticate requests to the service.
    * @param options - Used to configure the Search Index client.
    */
-  constructor(endpoint: string, credential: KeyCredential, options: SearchIndexClientOptions = {}) {
+  constructor(
+    endpoint: string,
+    credential: KeyCredential | TokenCredential,
+    options: SearchIndexClientOptions = {}
+  ) {
     this.endpoint = endpoint;
     this.credential = credential;
     this.options = options;
@@ -127,16 +141,21 @@ export class SearchIndexClient {
             "OData-MaxVersion",
             "OData-Version",
             "Prefer",
-            "throttle-reason"
-          ]
-        }
-      }
+            "throttle-reason",
+          ],
+        },
+      },
     };
 
-    const pipeline = createPipelineFromOptions(
-      internalPipelineOptions,
-      createSearchApiKeyCredentialPolicy(credential)
-    );
+    const scope: string = options.audience
+      ? `${options.audience}/.default`
+      : `${KnownSearchAudience.AzurePublicCloud}/.default`;
+
+    const requestPolicyFactory: RequestPolicyFactory = isTokenCredential(credential)
+      ? bearerTokenAuthenticationPolicy(credential, scope)
+      : createSearchApiKeyCredentialPolicy(credential);
+
+    const pipeline = createPipelineFromOptions(internalPipelineOptions, requestPolicyFactory);
 
     if (Array.isArray(pipeline.requestPolicyFactories)) {
       pipeline.requestPolicyFactories.unshift(odataMetadataPolicy("minimal"));
@@ -145,7 +164,7 @@ export class SearchIndexClient {
     let apiVersion = this.apiVersion;
 
     if (options.apiVersion) {
-      if (!["2020-06-30-Preview", "2020-06-30"].includes(options.apiVersion)) {
+      if (!["2020-06-30", "2021-04-30-Preview"].includes(options.apiVersion)) {
         throw new Error(`Invalid Api Version: ${options.apiVersion}`);
       }
       apiVersion = options.apiVersion;
@@ -167,7 +186,7 @@ export class SearchIndexClient {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -199,7 +218,7 @@ export class SearchIndexClient {
       },
       byPage: () => {
         return this.listIndexesPage(options);
-      }
+      },
     };
   }
 
@@ -210,14 +229,14 @@ export class SearchIndexClient {
     try {
       const result = await this.client.indexes.list({
         ...operationOptionsToRequestOptionsBase(updatedOptions),
-        select: "name"
+        select: "name",
       });
       const mapped = result.indexes.map((idx) => idx.name);
       yield mapped;
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -249,7 +268,7 @@ export class SearchIndexClient {
       },
       byPage: () => {
         return this.listIndexesNamesPage(options);
-      }
+      },
     };
   }
 
@@ -267,7 +286,7 @@ export class SearchIndexClient {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -284,13 +303,13 @@ export class SearchIndexClient {
     try {
       const result = await this.client.synonymMaps.list({
         ...operationOptionsToRequestOptionsBase(updatedOptions),
-        select: "name"
+        select: "name",
       });
       return result.synonymMaps.map((sm) => sm.name);
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -314,7 +333,7 @@ export class SearchIndexClient {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -341,7 +360,7 @@ export class SearchIndexClient {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -368,7 +387,7 @@ export class SearchIndexClient {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -395,7 +414,7 @@ export class SearchIndexClient {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -421,14 +440,14 @@ export class SearchIndexClient {
         utils.publicIndexToGeneratedIndex(index),
         {
           ...operationOptionsToRequestOptionsBase(updatedOptions),
-          ifMatch: etag
+          ifMatch: etag,
         }
       );
       return utils.generatedIndexToPublicIndex(result);
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -457,14 +476,14 @@ export class SearchIndexClient {
         utils.publicSynonymMapToGeneratedSynonymMap(synonymMap),
         {
           ...operationOptionsToRequestOptionsBase(updatedOptions),
-          ifMatch: etag
+          ifMatch: etag,
         }
       );
       return utils.generatedSynonymMapToPublicSynonymMap(result);
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -489,12 +508,12 @@ export class SearchIndexClient {
 
       await this.client.indexes.delete(indexName, {
         ...operationOptionsToRequestOptionsBase(updatedOptions),
-        ifMatch: etag
+        ifMatch: etag,
       });
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -523,12 +542,12 @@ export class SearchIndexClient {
 
       await this.client.synonymMaps.delete(synonymMapName, {
         ...operationOptionsToRequestOptionsBase(updatedOptions),
-        ifMatch: etag
+        ifMatch: etag,
       });
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -556,7 +575,7 @@ export class SearchIndexClient {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -580,7 +599,7 @@ export class SearchIndexClient {
         {
           ...restOptions,
           analyzer: restOptions.analyzerName,
-          tokenizer: restOptions.tokenizerName
+          tokenizer: restOptions.tokenizerName,
         },
         operationOptionsToRequestOptionsBase(updatedOptions)
       );
@@ -588,7 +607,7 @@ export class SearchIndexClient {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {
@@ -612,7 +631,7 @@ export class SearchIndexClient {
     } catch (e) {
       span.setStatus({
         code: SpanStatusCode.ERROR,
-        message: e.message
+        message: e.message,
       });
       throw e;
     } finally {