+++ sdk/search/search-documents/src/serviceUtils.ts
@@ -46,13 +46,12 @@ import {
   PatternAnalyzer as GeneratedPatternAnalyzer,
   CustomAnalyzer,
   PatternTokenizer,
-  LexicalNormalizerName
+  LexicalNormalizerName,
 } from "./generated/service/models";
 import {
   LexicalAnalyzer,
   CharFilter,
   CognitiveServicesAccount,
-  ComplexField,
   SearchField,
   SearchIndex,
   isComplexField,
@@ -70,12 +69,12 @@ import {
   SimilarityAlgorithm,
   SearchResourceEncryptionKey,
   PatternAnalyzer,
-  LexicalNormalizer
+  LexicalNormalizer,
 } from "./serviceModels";
 import { SuggestDocumentsResult, SuggestResult, SearchResult } from "./indexModels";
 import {
   SuggestDocumentsResult as GeneratedSuggestDocumentsResult,
-  SearchResult as GeneratedSearchResult
+  SearchResult as GeneratedSearchResult,
 } from "./generated/data/models";
 
 export function convertSkillsToPublic(skills: SearchIndexerSkillUnion[]): SearchIndexerSkill[] {
@@ -189,13 +188,13 @@ export function convertAnalyzersToGenerated(
       case "#Microsoft.Azure.Search.PatternAnalyzer":
         result.push({
           ...analyzer,
-          flags: analyzer.flags ? analyzer.flags.join("|") : undefined
+          flags: analyzer.flags ? analyzer.flags.join("|") : undefined,
         });
         break;
       case "#Microsoft.Azure.Search.CustomAnalyzer":
         result.push({
           ...analyzer,
-          tokenizerName: analyzer.tokenizerName
+          tokenizerName: analyzer.tokenizerName,
         });
         break;
     }
@@ -224,13 +223,13 @@ export function convertAnalyzersToPublic(
           ...analyzer,
           flags: (analyzer as GeneratedPatternAnalyzer).flags
             ? ((analyzer as GeneratedPatternAnalyzer).flags!.split("|") as RegexFlags[])
-            : undefined
+            : undefined,
         } as PatternAnalyzer);
         break;
       case "#Microsoft.Azure.Search.CustomAnalyzer":
         result.push({
           ...analyzer,
-          tokenizerName: (analyzer as CustomAnalyzer).tokenizerName
+          tokenizerName: (analyzer as CustomAnalyzer).tokenizerName,
         } as CustomAnalyzer);
         break;
     }
@@ -246,13 +245,17 @@ export function convertFieldsToPublic(fields: GeneratedSearchField[]): SearchFie
   return fields.map<SearchField>((field) => {
     let result: SearchField;
     if (field.type === "Collection(Edm.ComplexType)" || field.type === "Edm.ComplexType") {
-      result = field as ComplexField;
+      return {
+        name: field.name,
+        type: field.type,
+        fields: convertFieldsToPublic(field.fields!),
+      };
     } else {
-      const anayzerName: LexicalAnalyzerName | undefined | null = field.analyzer;
+      const analyzerName: LexicalAnalyzerName | undefined | null = field.analyzer;
       const searchAnalyzerName: LexicalAnalyzerName | undefined | null = field.searchAnalyzer;
       const indexAnalyzerName: LexicalAnalyzerName | undefined | null = field.indexAnalyzer;
       const synonymMapNames: string[] | undefined = field.synonymMaps;
-      const normalizerNames: LexicalNormalizerName | undefined | null = field.normalizer;
+      const normalizerName: LexicalNormalizerName | undefined | null = field.normalizer;
 
       const { retrievable, ...restField } = field;
       const hidden = typeof retrievable === "boolean" ? !retrievable : retrievable;
@@ -260,11 +263,11 @@ export function convertFieldsToPublic(fields: GeneratedSearchField[]): SearchFie
       result = {
         ...restField,
         hidden,
-        anayzerName,
+        analyzerName,
         searchAnalyzerName,
         indexAnalyzerName,
         synonymMapNames,
-        normalizerNames
+        normalizerName,
       } as SimpleField;
     }
     return result;
@@ -274,7 +277,11 @@ export function convertFieldsToPublic(fields: GeneratedSearchField[]): SearchFie
 export function convertFieldsToGenerated(fields: SearchField[]): GeneratedSearchField[] {
   return fields.map<GeneratedSearchField>((field) => {
     if (isComplexField(field)) {
-      return field;
+      return {
+        name: field.name,
+        type: field.type,
+        fields: convertFieldsToGenerated(field.fields),
+      };
     } else {
       const { hidden, ...restField } = field;
       const retrievable = typeof hidden === "boolean" ? !hidden : hidden;
@@ -290,7 +297,7 @@ export function convertFieldsToGenerated(fields: SearchField[]): GeneratedSearch
         searchAnalyzer: field.searchAnalyzerName,
         indexAnalyzer: field.indexAnalyzerName,
         synonymMaps: field.synonymMapNames,
-        normalizer: field.normalizerName
+        normalizer: field.normalizerName,
       };
     }
   });
@@ -308,7 +315,7 @@ export function convertTokenizersToGenerated(
     if (tokenizer.odatatype === "#Microsoft.Azure.Search.PatternTokenizer") {
       result.push({
         ...tokenizer,
-        flags: tokenizer.flags ? tokenizer.flags.join("|") : undefined
+        flags: tokenizer.flags ? tokenizer.flags.join("|") : undefined,
       });
     } else {
       result.push(tokenizer);
@@ -331,7 +338,7 @@ export function convertTokenizersToPublic(
         ...tokenizer,
         flags: (tokenizer as PatternTokenizer).flags
           ? ((tokenizer as PatternTokenizer).flags!.split("|") as RegexFlags[])
-          : undefined
+          : undefined,
       });
     } else {
       result.push(tokenizer);
@@ -376,9 +383,9 @@ export function extractOperationOptions<T extends OperationOptions>(
     operationOptions: {
       abortSignal,
       requestOptions,
-      tracingOptions
+      tracingOptions,
     },
-    restOptions
+    restOptions,
   };
 }
 
@@ -392,7 +399,7 @@ export function convertEncryptionKeyToPublic(
   const result: SearchResourceEncryptionKey = {
     keyName: encryptionKey.keyName,
     keyVersion: encryptionKey.keyVersion,
-    vaultUrl: encryptionKey.vaultUri
+    vaultUrl: encryptionKey.vaultUri,
   };
 
   if (encryptionKey.accessCredentials) {
@@ -413,13 +420,13 @@ export function convertEncryptionKeyToGenerated(
   const result: GeneratedSearchResourceEncryptionKey = {
     keyName: encryptionKey.keyName,
     keyVersion: encryptionKey.keyVersion,
-    vaultUri: encryptionKey.vaultUrl
+    vaultUri: encryptionKey.vaultUrl,
   };
 
   if (encryptionKey.applicationId) {
     result.accessCredentials = {
       applicationId: encryptionKey.applicationId,
-      applicationSecret: encryptionKey.applicationSecret
+      applicationSecret: encryptionKey.applicationSecret,
     };
   }
 
@@ -441,7 +448,7 @@ export function generatedIndexToPublicIndex(generatedIndex: GeneratedSearchIndex
     normalizers: generatedIndex.normalizers as LexicalNormalizer[],
     scoringProfiles: generatedIndex.scoringProfiles as ScoringProfile[],
     fields: convertFieldsToPublic(generatedIndex.fields),
-    similarity: convertSimilarityToPublic(generatedIndex.similarity)
+    similarity: convertSimilarityToPublic(generatedIndex.similarity),
   };
 }
 
@@ -451,14 +458,14 @@ export function generatedSearchResultToPublicSearchResult<T>(
   const returnValues: SearchResult<T>[] = results.map<SearchResult<T>>((result) => {
     const { _score, _highlights, rerankerScore, captions, ...restProps } = result;
     const doc: { [key: string]: any } = {
-      ...restProps
+      ...restProps,
     };
     const obj = {
       score: _score,
       highlights: _highlights,
       rerankerScore,
       captions,
-      document: doc
+      document: doc,
     };
     return obj as SearchResult<T>;
   });
@@ -472,12 +479,12 @@ export function generatedSuggestDocumentsResultToPublicSuggestDocumentsResult<T>
     const { _text, ...restProps } = element;
 
     const doc: { [key: string]: any } = {
-      ...restProps
+      ...restProps,
     };
 
     const obj = {
       text: _text,
-      document: doc
+      document: doc,
     };
 
     return obj as SuggestResult<T>;
@@ -485,7 +492,7 @@ export function generatedSuggestDocumentsResultToPublicSuggestDocumentsResult<T>
 
   const result: SuggestDocumentsResult<T> = {
     results: results,
-    coverage: searchDocumentsResult.coverage
+    coverage: searchDocumentsResult.coverage,
   };
 
   return result;
@@ -506,7 +513,7 @@ export function publicIndexToGeneratedIndex(index: SearchIndex): GeneratedSearch
     analyzers: convertAnalyzersToGenerated(index.analyzers),
     tokenizers: convertTokenizersToGenerated(index.tokenizers),
     fields: convertFieldsToGenerated(index.fields),
-    similarity: convertSimilarityToGenerated(index.similarity)
+    similarity: convertSimilarityToGenerated(index.similarity),
   };
 }
 
@@ -522,7 +529,7 @@ export function generatedSkillsetToPublicSkillset(
     ),
     knowledgeStore: generatedSkillset.knowledgeStore,
     etag: generatedSkillset.etag,
-    encryptionKey: convertEncryptionKeyToPublic(generatedSkillset.encryptionKey)
+    encryptionKey: convertEncryptionKeyToPublic(generatedSkillset.encryptionKey),
   };
 }
 
@@ -538,7 +545,7 @@ export function publicSkillsetToGeneratedSkillset(
       skillset.cognitiveServicesAccount
     ),
     knowledgeStore: skillset.knowledgeStore,
-    encryptionKey: convertEncryptionKeyToGenerated(skillset.encryptionKey)
+    encryptionKey: convertEncryptionKeyToGenerated(skillset.encryptionKey),
   };
 }
 
@@ -547,7 +554,7 @@ export function generatedSynonymMapToPublicSynonymMap(synonymMap: GeneratedSynon
     name: synonymMap.name,
     encryptionKey: convertEncryptionKeyToPublic(synonymMap.encryptionKey),
     etag: synonymMap.etag,
-    synonyms: []
+    synonyms: [],
   };
 
   if (synonymMap.synonyms) {
@@ -563,7 +570,7 @@ export function publicSynonymMapToGeneratedSynonymMap(synonymMap: SynonymMap): G
     format: "solr",
     encryptionKey: convertEncryptionKeyToGenerated(synonymMap.encryptionKey),
     etag: synonymMap.etag,
-    synonyms: synonymMap.synonyms.join("\n")
+    synonyms: synonymMap.synonyms.join("\n"),
   };
 
   result.encryptionKey = convertEncryptionKeyToGenerated(synonymMap.encryptionKey);
@@ -576,7 +583,7 @@ export function publicSearchIndexerToGeneratedSearchIndexer(
 ): GeneratedSearchIndexer {
   return {
     ...indexer,
-    encryptionKey: convertEncryptionKeyToGenerated(indexer.encryptionKey)
+    encryptionKey: convertEncryptionKeyToGenerated(indexer.encryptionKey),
   };
 }
 
@@ -585,7 +592,7 @@ export function generatedSearchIndexerToPublicSearchIndexer(
 ): SearchIndexer {
   return {
     ...indexer,
-    encryptionKey: convertEncryptionKeyToPublic(indexer.encryptionKey)
+    encryptionKey: convertEncryptionKeyToPublic(indexer.encryptionKey),
   };
 }
 
@@ -597,13 +604,13 @@ export function publicDataSourceToGeneratedDataSource(
     description: dataSource.description,
     type: dataSource.type,
     credentials: {
-      connectionString: dataSource.connectionString
+      connectionString: dataSource.connectionString,
     },
     container: dataSource.container,
     etag: dataSource.etag,
     dataChangeDetectionPolicy: dataSource.dataChangeDetectionPolicy,
     dataDeletionDetectionPolicy: dataSource.dataDeletionDetectionPolicy,
-    encryptionKey: convertEncryptionKeyToGenerated(dataSource.encryptionKey)
+    encryptionKey: convertEncryptionKeyToGenerated(dataSource.encryptionKey),
   };
 }
 
@@ -623,7 +630,7 @@ export function generatedDataSourceToPublicDataSource(
     dataDeletionDetectionPolicy: convertDataDeletionDetectionPolicyToPublic(
       dataSource.dataDeletionDetectionPolicy
     ),
-    encryptionKey: convertEncryptionKeyToPublic(dataSource.encryptionKey)
+    encryptionKey: convertEncryptionKeyToPublic(dataSource.encryptionKey),
   };
 }